refactor(dsh-mcp-manager): 阶段3连接域逻辑收敛——connection/{orchestrator,runtime} 成形（interface.ts 门面）
Summary:
- 新建 src/connection/{orchestrator,runtime} + 三级 interface.ts（connection/
  orchestrator/ runtime/）：连接域唯一对外引用面（D10）
- 重连策略解析（ReconnectPolicy/RECONNECT_DEFAULTS/resolveReconnect）自
  supervisor.ts 迁入 connection/runtime/reconnect.ts，成为两路径（supervisor +
  中间层）退避/预算口径的单一事实源（B18 收敛基础）
- 工具名展示投影 stripMcpPrefix 自 manager.ts 迁入 connection/orchestrator/
  tool-names.ts（#382 F4 裸名展示口径单一事实源）
- supervisor.ts / manager.ts / index.ts 的连接域引用改走 connection/interface.ts；
  lib re-export 面不变（index.ts 顶层导出零变化）
- 零行为变更：build + typecheck + verify-dir-imports + 全量 smoke 绿

Design decisions:
- 阶段 3 只做「文件内重组」：类主体（McpManager/ConnectionSupervisor/
  McpMiddleware）仍在 src 根目录对应文件，阶段 6 集中搬移；本阶段仅迁出
  无 this 依赖的纯逻辑段，杜绝拆类带来的测试面碎裂
- 连接域门面不代理 middleware.ts 的 workspace/catalog/pipeline 符号（那些域
  阶段 4/5 归位），避免污染域边界；index.ts 连接域 re-export 仅重连策略改走
  门面，其余保持现状
- verify-dir-imports 只查 src 直接子目录层，connection/runtime/ 等二级目录
  暂不受静态强制；仍按 D10 精神逐目录补 interface.ts，阶段 6 收紧不红

Related issue: #664
